fix: correct SP calculation — balanced diet ratio, float precision, zero-cal filter
Balanced diet ratio was excluding zero nutrients from min(), making diets
appear more balanced than the game calculates (e.g. Pumpkin with fat=0
gave ratio=0.2 instead of 0, SP 17.75 instead of game's 16). Now uses
min(all nutrients) including zeros, matching Eco's BalancedDietMult.

Also fixes C# mod truncating float nutrients/calories to int (losing
precision from Eco API), and filters out zero-calorie items (seeds,
spores) that contributed nothing but consumed plan slots.

+++ mod/EcoDietMod/Algorithm/MealPlanner.cs
@@ -28,7 +28,7 @@ public static class MealPlanner
         Dictionary<FoodCandidate, int> available,
         List<string> cravings,
         int cravingsSatisfied,
-        int remainingCalories,
+        float remainingCalories,
         PlannerConfig config,
         float serverMult = 1f,
         float dinnerPartyMult = 1f)
@@ -90,7 +90,7 @@ public static class MealPlanner
     /// Consume one unit of food, recompute SP/bonuses, append to plan items.
     /// Returns updated (currentSp, remainingCalories, cravingsSatisfied, varietyCount).
     /// </summary>
-    private static (float CurrentSp, int RemainingCalories, int CravingsSatisfied, int VarietyCount)
+    private static (float CurrentSp, float RemainingCalories, int CravingsSatisfied, int VarietyCount)
         ApplyBite(
             Dictionary<FoodCandidate, int> stomach,
             Dictionary<FoodCandidate, int> available,
@@ -98,7 +98,7 @@ public static class MealPlanner
             List<string> cravings,
             List<MealPlanItem> items,
             float currentSp,
-            int remainingCalories,
+            float remainingCalories,
             int cravingsSatisfied,
             int varietyCountNow,
             PlannerConfig config,
